add cludinary upload image for profile

+++ src/controller/profile.controller.ts
@@ -7,6 +7,8 @@ import {
   updateDataProfileService,
 } from "../service/profile.service";
 import { createProfileSchemas } from "../utils/schemas/profile.schemas";
+import { v2 as cloudinary, UploadApiResponse } from "cloudinary";
+import fs from "fs";
 
 export const getDataProfileController = async (
   req: Request,
@@ -30,10 +32,20 @@ export const createDataProfileController = async (
 ) => {
   try {
     const userId = (req as any).user.id;
-    const body = req.body;
+
+    let uploadResult: UploadApiResponse = {} as UploadApiResponse;
+
+    if (req.file) {
+      uploadResult = await cloudinary.uploader.upload(req.file.path || "");
+      fs.unlinkSync(req.file.path);
+    }
+
+    const body = { ...req.body, images: uploadResult.secure_url ?? undefined };
 
     const { error, value } = createProfileSchemas.validate(body);
+
     if (error) {
+      await cloudinary.uploader.destroy(uploadResult.public_id);
       res.status(400).json({ error: error.details[0].message });
       return;
     }
@@ -53,18 +65,38 @@ export const updateDataProfileController = async (
 ) => {
   try {
     const { id } = req.params;
-    const body = req.body;
+    let uploadResult: UploadApiResponse = {} as UploadApiResponse;
+
+    const checkData = await getDataProfileFindByIdService(id);
+    if (!checkData) {
+      res.status(404).json({ message: "Data Not found" });
+      return;
+    }
+
+    const getDataById = await getDataProfileFindByIdService(id);
+    const url = getDataById?.images;
+    const fileName = url?.substring(url.lastIndexOf("/") + 1).split(".")[0];
+
+    if (req.file) {
+      uploadResult = await cloudinary.uploader.upload(req.file.path || "");
+      fs.unlinkSync(req.file.path);
+    }
+
+    const body = { ...req.body, images: uploadResult.secure_url ?? undefined };
 
     const { error, value } = createProfileSchemas.validate(body);
-
     if (error) {
+      await cloudinary.uploader.destroy(uploadResult.secure_url);
       res.status(400).json({ error: error.details[0].message });
       return;
     }
 
-    const data = await updateDataProfileService(id, value);
-
-    res.status(200).json({ message: "success", data });
+    if (value) {
+      await cloudinary.uploader.destroy(fileName || "");
+      const data = await updateDataProfileService(id, value);
+      res.status(200).json({ message: "success", data });
+      return;
+    }
   } catch (error) {
     next(error);
   }
@@ -77,6 +109,11 @@ export const deleteDataProfileController = async (
 ) => {
   try {
     const { id } = req.params;
+
+    const getDataById = await getDataProfileFindByIdService(id);
+    const url = getDataById?.images;
+    const fileName = url?.substring(url.lastIndexOf("/") + 1).split(".")[0];
+
     const checkData = await getDataProfileFindByIdService(id);
 
     if (!checkData) {
@@ -84,9 +121,13 @@ export const deleteDataProfileController = async (
       return;
     }
 
-    const data = await deleteDataProfileService(id);
+    const deleteImage = await cloudinary.uploader.destroy(fileName || "");
 
-    res.status(200).json({ message: "success", data });
+    if (deleteImage) {
+      const data = await deleteDataProfileService(id);
+      res.status(200).json({ message: "success", data });
+      return;
+    }
   } catch (error) {
     next(error);
   }
